pr feedback, break index.ts into files
